Fix warnings

+++ tests/microbenchs/tasks_size_overhead.c
@@ -70,7 +70,7 @@ void func(void *descr[], void *arg)
 		double tv2 = starpu_timing_now();
 		usec = tv2 - tv1;
 	}
-	while (usec < n);
+	while (usec < (long) n);
 }
 
 double cost_function(struct starpu_task *t, struct starpu_perfmodel_arch *a, unsigned i)
